+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Azure Blob Storage support.
 - `IMGPROXY_STRIP_COLOR_PROFILE` config and [strip_color_profile](https://docs.imgproxy.net/#/generating_the_url_advanced?id=strip-color-profile) processing option.
 - `IMGPROXY_FORMAT_QUALITY` config.
+- `IMGPROXY_AUTO_ROTATE` config and [auto_rotate](https://docs.imgproxy.net/#/generating_the_url_advanced?id=auto-rotate) processing option.
 - (pro) Remove Adobe Illustrator garbage from SVGs.
 
 ### Changed